requirements.txt
# داده و پردازش
pandas==2.1.0
numpy==1.24.0
ta-lib==0.4.24
yfinance==0.2.28
ccxt==4.1.0
alpha-vantage==2.3.1
requests==2.31.0

# یادگیری ماشین
scikit-learn==1.3.0
xgboost==1.7.5
lightgbm==4.1.0
statsmodels==0.14.0

# یادگیری عمیق
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1

# بکتست و تریدینگ
backtrader==1.9.78.123
vectorbt==0.25.6
backtesting.py==0.3.3
pyfolio==0.9.2

# ویژوال‌سازی
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
mplfinance==0.12.9b7

# دپلوی و دشبورد
streamlit==1.27.0
fastapi==0.104.0
uvicorn==0.24.0

# ابزارها
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1
jupyter==1.0.0